requirements.txt
hikari[server]
hikari-arc[rest]
fastapi
uvicorn
python-dotenv